+++ packages/graphql-hooks/src/utils.js
@@ -1,0 +1,7 @@
+/**
+ * Pipe with support for async functions
+ * @param {array} array of functions
+ * @returns single function
+ */
+export const pipeP = fns => arg =>
+  fns.reduce((p, f) => p.then(f), Promise.resolve(arg))