+++ examples/subscription/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.6.0](https://github.com/nearform/graphql-hooks/compare/subscription@2.5.0...subscription@2.6.0) (2022-03-25)
+
+**Note:** Version bump only for package subscription
+
+
+
+
+
+# [2.5.0](https://github.com/nearform/graphql-hooks/compare/subscription@2.4.1...subscription@2.5.0) (2022-03-25)
+
+**Note:** Version bump only for package subscription
+
+
+
+
+
+## [2.4.1](https://github.com/nearform/graphql-hooks/compare/subscription@2.4.0...subscription@2.4.1) (2022-03-11)
+
+**Note:** Version bump only for package subscription
+
+
+
+
+
 # [2.4.0](https://github.com/nearform/graphql-hooks/compare/subscription@2.3.0...subscription@2.4.0) (2022-02-10)
 
 **Note:** Version bump only for package subscription