+++ examples/typescript/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.7.0](https://github.com/nearform/graphql-hooks/compare/typescript-example@2.6.0...typescript-example@2.7.0) (2022-03-25)
+
+**Note:** Version bump only for package typescript-example
+
+
+
+
+
+# [2.6.0](https://github.com/nearform/graphql-hooks/compare/typescript-example@2.5.1...typescript-example@2.6.0) (2022-03-25)
+
+**Note:** Version bump only for package typescript-example
+
+
+
+
+
+## [2.5.1](https://github.com/nearform/graphql-hooks/compare/typescript-example@2.5.0...typescript-example@2.5.1) (2022-03-11)
+
+**Note:** Version bump only for package typescript-example
+
+
+
+
+
 # [2.5.0](https://github.com/nearform/graphql-hooks/compare/typescript-example@2.4.0...typescript-example@2.5.0) (2022-02-10)
 
 **Note:** Version bump only for package typescript-example