+++ packages/graphql-hooks/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.9.0](https://github.com/nearform/graphql-hooks/compare/graphql-hooks@5.8.0...graphql-hooks@5.9.0) (2022-03-25)
+
+
+### Features
+
+* local mock client for graphql-hooks ([#797](https://github.com/nearform/graphql-hooks/issues/797)) ([23a9526](https://github.com/nearform/graphql-hooks/commit/23a95264c9c70cb6745070ba0342bbb1661d7e22))
+
+
+
+
+
+# [5.8.0](https://github.com/nearform/graphql-hooks/compare/graphql-hooks@5.7.1...graphql-hooks@5.8.0) (2022-03-25)
+
+**Note:** Version bump only for package graphql-hooks
+
+
+
+
+
+## [5.7.1](https://github.com/nearform/graphql-hooks/compare/graphql-hooks@5.7.0...graphql-hooks@5.7.1) (2022-03-11)
+
+
+### Bug Fixes
+
+* dont overwrite the cache for ssrMode ([#787](https://github.com/nearform/graphql-hooks/issues/787)) ([754e191](https://github.com/nearform/graphql-hooks/commit/754e191822cd4b0cbff78f3701f7e35f6d6451af))
+
+
+### Features
+
+* **typescript:** adds cacheKey to UseClientRequestResult type ([#793](https://github.com/nearform/graphql-hooks/issues/793)) ([da796ce](https://github.com/nearform/graphql-hooks/commit/da796ceb7d0877f9d162ee4a7e25c7a7dc6889a0))
+
+
+
+
+
 # [5.7.0](https://github.com/nearform/graphql-hooks/compare/graphql-hooks@5.6.0...graphql-hooks@5.7.0) (2022-02-10)